--- a/src/webgpu/shader/execution/expression/expression.ts
+++ b/src/webgpu/shader/execution/expression/expression.ts
@@ -353,6 +353,20 @@ export async function run(
     }
   };
 
+  const processBatch = async (batchCases: CaseList) => {
+    const checkBatch = await submitBatch(
+      t,
+      shaderBuilder,
+      parameterTypes,
+      resultType,
+      batchCases,
+      cfg.inputSource,
+      pipelineCache
+    );
+    checkBatch();
+    void t.queue.onSubmittedWorkDone().finally(batchFinishedCallback);
+  }
+
   for (let i = 0; i < cases.length; i += casesPerBatch) {
     const batchCases = cases.slice(i, Math.min(i + casesPerBatch, cases.length));
 
@@ -365,17 +379,7 @@ export async function run(
     }
     batchesInFlight += 1;
 
-    const checkBatch = submitBatch(
-      t,
-      shaderBuilder,
-      parameterTypes,
-      resultType,
-      batchCases,
-      cfg.inputSource,
-      pipelineCache
-    );
-    checkBatch();
-    void t.queue.onSubmittedWorkDone().finally(batchFinishedCallback);
+    processBatch(batchCases);
   }
 }
 
@@ -391,7 +395,7 @@ export async function run(
  * @param pipelineCache the cache of compute pipelines, shared between batches
  * @returns a function that checks the results are as expected
  */
-function submitBatch(
+async function submitBatch(
   t: GPUTest,
   shaderBuilder: ShaderBuilder,
   parameterTypes: Array<Type>,
@@ -399,7 +403,7 @@ function submitBatch(
   cases: CaseList,
   inputSource: InputSource,
   pipelineCache: PipelineCache
-): () => void {
+): Promise<() => void> {
   // Construct a buffer to hold the results of the expression tests
   const outputBufferSize = cases.length * valueStride(resultType);
   const outputBuffer = t.device.createBuffer({
@@ -407,7 +411,7 @@ function submitBatch(
     usage: GPUBufferUsage.COPY_SRC | GPUBufferUsage.COPY_DST | GPUBufferUsage.STORAGE,
   });
 
-  const [pipeline, group] = buildPipeline(
+  const [pipeline, group] = await buildPipeline(
     t,
     shaderBuilder,
     parameterTypes,
@@ -1003,7 +1007,7 @@ ${body}
  * @param outputBuffer the buffer that will hold the output values of the tests
  * @param pipelineCache the cache of compute pipelines, shared between batches
  */
-function buildPipeline(
+async function buildPipeline(
   t: GPUTest,
   shaderBuilder: ShaderBuilder,
   parameterTypes: Array<Type>,
@@ -1012,7 +1016,7 @@ function buildPipeline(
   inputSource: InputSource,
   outputBuffer: GPUBuffer,
   pipelineCache: PipelineCache
-): [GPUComputePipeline, GPUBindGroup] {
+): Promise<[GPUComputePipeline, GPUBindGroup]> {
   cases.forEach(c => {
     const inputTypes = c.input instanceof Array ? c.input.map(i => i.type) : [c.input.type];
     if (!objectEquals(inputTypes, parameterTypes)) {
@@ -1032,7 +1036,7 @@ function buildPipeline(
       const module = t.device.createShaderModule({ code: source });
 
       // build the pipeline
-      const pipeline = t.device.createComputePipeline({
+      const pipeline = await t.device.createComputePipelineAsync({
         layout: 'auto',
         compute: { module, entryPoint: 'main' },
       });