Some of these tests are among the longest running that we have, so hopefully this has a dramatic impact on runtimes. Profiling shows that almost all the time is being spent waiting for pipeline creation. By using createComputePipelineAsync, however, we ensure that multiple shaders in a batch can be compiling at once, which drops test run times significantly.

For example, the `webgpu:shader,execution,expression,binary,af_matrix_subtraction:matrix:inputSource="const";cols=4;rows=4` test, one of the longest running in the entire CTS, was taking 5 minutes to run on my Windows machine. With this patch it completes in ~~46 seconds~~ 80 seconds on the same device.

In general I'm seeing **speedups of ~~5x-6x~~** (sorry, it's only **4x**. See comments below) on the tests that I've tried.

To demonstrate why, we can look at this chunk of a trace from the shorter `webgpu:shader,execution,expression,binary,af_matrix_subtraction:matrix:inputSource="const";cols=2;rows=2` test, which was previously taking 8 seconds to run.

<img width="840" alt="Screenshot 2023-11-01 104727" src="https://github.com/gpuweb/cts/assets/805273/999ecfb0-15fc-4277-89ba-222ded69932d">

The blue chunks are all pipeline compiles, and you can see they dominate the runtime and are being performed serially. With this patch it completes in about 2 seconds and the tracing produces results like this:

<img width="553" alt="Screenshot 2023-11-01 105615" src="https://github.com/gpuweb/cts/assets/805273/aa18626c-3031-4331-adf8-99fe0e60fda1">

Same amount of blue, but now more parallel!

<hr>

**Requirements for PR author:**

- [x] All missing test coverage is tracked with "TODO" or `.unimplemented()`.
- [x] New helpers are `/** documented */` and new helper files are found in `helper_index.txt`.
- [x] Test behaves as expected in a WebGPU implementation. (If not passing, explain above.)

**Requirements for [reviewer sign-off](https://github.com/gpuweb/cts/blob/main/docs/reviews.md):**

- [ ] Tests are properly located in the test tree.
- [ ] [Test descriptions](https://github.com/gpuweb/cts/blob/main/docs/intro/plans.md) allow a reader to "read only the test plans and evaluate coverage completeness", and accurately reflect the test code.
- [ ] Tests provide complete coverage (including validation control cases). **Missing coverage MUST be covered by TODOs.**
- [ ] Helpers and types promote readability and maintainability.

When landing this PR, be sure to make any necessary issue status updates.